requirements.txt
annotated-types==0.6.0
anyio==3.7.1
appscript==1.2.5
certifi==2024.2.2
distro==1.9.0
docx2pdf==0.1.8
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
lxml==4.9.4
openai==1.14.1
pydantic==2.6.4
pydantic_core==2.16.3
python-docx==1.1.0
python-dotenv==1.0.1
PyYAML==6.0.1
sniffio==1.3.1
tqdm==4.66.2
typing_extensions==4.10.0
